+++ packages/web-components/drawer/src/components/index.ts
@@ -1,0 +1,4 @@
+export * from "./drawer-content.js";
+export * from "./drawer-trigger.js";
+export * from "./drawer-backdrop.js";
+export * from "./drawer-context.js";