+++ packages/web-components/utils/src/isCustomElement.ts
@@ -1,5 +1,5 @@
 import type { ICustomElement } from "component-register";
-export default function isCustomElement(
+export function isCustomElement(
   element: any,
 ): element is ICustomElement & HTMLElement {
   return Reflect.has(element, "__initialized");