+++ packages/web-components/utils/src/toHyphenated.ts
@@ -1,4 +1,4 @@
-export default function toHyphenated(str: string): string {
+export function toHyphenated(str: string): string {
   return str
     .replace(/([A-Z])/g, "-$1") // Insert hyphen before uppercase letters
     .toLowerCase(); // Convert the entire string to lowercase